Added get request for employee details and orders

+++ App.BusinessLogic/Helper/CustomDate.cs
@@ -14,7 +14,7 @@ namespace App.BusinessLogic.Helper
         public static void CustomAction(CleaningServiceOrder input, CustomMappingOutput output)
         {
             var customDate = Convert.ToDateTime(input.DateTime);
-            output.CustomDateHour = (float)(((customDate.Year * 10 + customDate.Month) * 10 + customDate.Day) * 10 + customDate.Hour) * 10;
+            output.CustomDateHour = (((customDate.Year * 10 + customDate.Month) * 10 + customDate.Day) * 10 + customDate.Hour) * 10;
         }
 
         public override Action<CleaningServiceOrder, CustomMappingOutput> GetMapping() => CustomAction;
